Adjust corert for shared SafeHandle and EventRegistrationToken
